chore rename highlight interfaces to types
