+++ src/Highlight/index.ts
@@ -1,4 +1,4 @@
-export { IHighlight, IHighlightCollection, IHighlightCollectionAppModel } from "./interfaces";
+export { IHighlight, IHighlightCollection, IHighlightCollectionAppModel } from "./types";
 export { HighlightContainerRuntimeFactory } from "./container";
 export { Highlight, HighlightCollection } from "./model";
 export { renderHighlightCollection } from "./view";
