+++ src/Highlight/types.ts
